--- oras/cli/pull.py
@@ -1,28 +1,0 @@
-__author__ = "Vanessa Sochat"
-__copyright__ = "Copyright The ORAS Authors."
-__license__ = "Apache-2.0"
-
-import oras.client
-from oras.logger import logger
-
-
-def main(args, parser, extra, subparser):
-    """
-    A wrapper around an oras client pull.
-    """
-    client = oras.client.OrasClient(insecure=args.insecure)
-    try:
-        if args.username and args.password:
-            client.set_basic_auth(args.username, args.password)
-        client.pull(
-            config_path=args.config,
-            allowed_media_type=args.allowed_media_type
-            if not args.allow_all_media_types
-            else [],
-            overwrite=not args.keep_old_files,
-            manifest_config_ref=args.manifest_config_ref,
-            outdir=args.output,
-            target=args.target,
-        )
-    except Exception as e:
-        logger.exit(str(e))
