--- oras/main/__init__.py
@@ -1,1 +1,0 @@
-from .login import login
